moved FlipBits to _gates.py
