+++ projectq/ops/__init__.py
@@ -35,4 +35,4 @@ from ._shortcuts import *
 from ._time_evolution import TimeEvolution
 from ._uniformly_controlled_rotation import (UniformlyControlledRy,
                                              UniformlyControlledRz)
-from ._state_prep import StatePreparation, FlipBits
+from ._state_prep import StatePreparation
