+++ projectq/ops/_gates_test.py
@@ -217,3 +217,18 @@ def test_barrier_gate():
     assert str(gate) == "Barrier"
     assert gate.get_inverse() == _gates.BarrierGate()
     assert isinstance(_gates.Barrier, _gates.BarrierGate)
+
+
+def test_flip_bits_equality_and_hash():
+    gate1 = _gates.FlipBits([1, 0, 0, 1])
+    gate2 = _gates.FlipBits([1, 0, 0, 1])
+    gate3 = _gates.FlipBits([0, 1, 0, 1])
+    assert gate1 == gate2
+    assert hash(gate1) == hash(gate2)
+    assert gate1 != gate3
+    assert gate1 != _gates.X
+
+
+def test_flip_bits_str():
+    gate1 = _gates.FlipBits([0, 1, 0])
+    assert str(gate1) == "FlipBits"
