+++ projectq/ops/_state_prep.py
@@ -13,7 +13,6 @@
 #   limitations under the License.
 
 from ._basics import BasicGate, SelfInverseGate
-from ._gates import XGate
 
 class StatePreparation(BasicGate):
     """
@@ -55,50 +54,3 @@ class StatePreparation(BasicGate):
 
     def __hash__(self):
         return hash("StatePreparation(" + str(self.final_state) + ")")
-
-class FlipBits(SelfInverseGate):
-    """
-    Gate for transforming qubits in state |0> to any computational basis state
-    """
-    def __init__(self, bits_to_flip):
-        """
-        Initialize FlipBits gate.
-
-        Example:
-            .. code-block:: python
-
-                qureg = eng.allocate_qureg(2)
-                FlipBits([0, 1]) | qureg
-
-        Note:
-            The amplitude of state k is final_state[k]. When the state k is
-            written in binary notation, then qureg[0] denotes the qubit
-            whose state corresponds to the least significant bit of k.
-
-        Args:
-            bits_to_flip(list[int]|list[bool]|str): array of 0/1, True/False or
-                                    string of 0/1 identifying the qubits to flip
-                                    of length len(qureg).
-        """
-        if isinstance(bits_to_flip, str):
-            self.bits_to_flip = list([ c != "0" for c in bits_to_flip])
-        else:
-            self.bits_to_flip = list(bits_to_flip)
-
-    def __str__(self):
-        return "FlipBits"
-
-    def __or__(self, qubits):
-        for qureg in self.make_tuple_of_qureg(qubits):
-            for i, qubit in enumerate(qureg):
-                if self.bits_to_flip[i]:
-                    XGate() | qubit
-
-    def __eq__(self, other):
-        if isinstance(other, self.__class__):
-            return self.bits_to_flip == other.bits_to_flip
-        else:
-            return False
-
-    def __hash__(self):
-        return hash("FlipBits(" + str(self.bits_to_flip) + ")")
